rename topics to channels
